requirements.txt
certifi==2026.1.4
influxdb-client==1.49.0
pymodbus==3.11.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
reactivex==4.1.0
setuptools==80.9.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
